requirements.txt
concurrent-log-handler==0.9.16
defusedxml==0.6.0
Django==2.2.12
django-hosts==4.0
djangorestframework==3.10.2
djangorestframework-xml==2.0.0
mysqlclient==1.4.4
portalocker==1.7.0
pytz==2020.1
sqlparse==0.3.1
uWSGI==2.0.18
